fixup! ASoC: SOF: fix range checks
Remove explicit cast as suggested by Marc Herbert

+++ sound/soc/sof/topology.c
@@ -1163,7 +1163,7 @@ static int sof_control_load_bytes(struct snd_soc_component *scomp,
 	/* init the get/put bytes data */
 	if (priv_size > max_size - sizeof(struct sof_ipc_ctrl_data)) {
 		dev_err(scomp->dev, "err: bytes data size %zu exceeds max %zu.\n",
-			priv_size, (size_t)max_size - sizeof(struct sof_ipc_ctrl_data));
+			priv_size, max_size - sizeof(struct sof_ipc_ctrl_data));
 		ret = -EINVAL;
 		goto out;
 	}
